# Assorted Refactorings Part 2