This is a continuation of #404, with items that Oleg had indicated as potentially risky due to performance implications as mentioned here: https://github.com/sangria-graphql/sangria/pull/404#pullrequestreview-162264728

I'd like to keep the changes around, so I have split this into a separate PR in hopes that we can come up with a plan to validate, or disprove Oleg's concerns.