@yanns 2020-12-11
Now that we have a module for benchmark, one could write a test for this part and compare before & after the change.