@nickhudkins 2020-12-31
Hey! It looks like maybe the reason Oleg had indicated the risk was that he experienced performance sensitivity in this area already.

### Master:
[info] Result "sangria.benchmarks.OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete":
[info]   33743.582 ±(99.9%) 669.221 ops/s [Average]
[info]   (min, avg, max) = (31797.042, 33743.582, 35039.683), stdev = 893.391
[info]   CI (99.9%): [33074.362, 34412.803] (assumes normal distribution)

### This Branch:
[info] Result "sangria.benchmarks.OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete":
[info]   34008.015 ±(99.9%) 255.573 ops/s [Average]
[info]   (min, avg, max) = (33177.557, 34008.015, 34704.334), stdev = 341.183
[info]   CI (99.9%): [33752.442, 34263.588] (assumes normal distribution)

Happy to let these benchmarks run longer, but after the first pass through (5 forks, 5 iteration warmup, 5 test runs), the results look promising that we're not introducing a performance regression by accepting this PR.

@yanns Happy to leave it up to you if you would like me to run longer benchmarks before accepting this is A.O.K.

@nightscape I actually wonder if your hunch about collection methods being more optimized may play out here in the long run!