@yanns 2020-12-31
From 669.221 ops/s to 255.573 ops/s is quite a performance degradation, right?