@nickhudkins 2020-12-31
Maybe I am reading it wrong, but my understanding of this is that it's 

33743.582 vs 34008.015 ? 

I'll read more about these results, but I believe the 669 and 255 are somehow related to the standard deviation, not the raw throughput.