@nickhudkins 2020-12-31
I'll go ahead and get a full run running *in the cloud* so I can let it sit