@nickhudkins 2021-01-06
### Master
```
Benchmark                                                            (size)     (testCase)  (validatorType)   Mode  Cnt         Score        Error  Units
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete       2            N/A              Old  thrpt   25     18271.535 ±     59.183  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete       2            N/A              New  thrpt   25     19278.953 ±    108.384  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete      10            N/A              Old  thrpt   25      5598.077 ±     61.633  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete      10            N/A              New  thrpt   25      6255.941 ±     15.719  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete     100            N/A              Old  thrpt   25       158.272 ±      1.055  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete     100            N/A              New  thrpt   25       698.410 ±      2.674  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag              2            N/A              Old  thrpt   25     70177.779 ±    392.781  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag              2            N/A              New  thrpt   25     42535.811 ±    253.822  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag             10            N/A              Old  thrpt   25      2887.106 ±     27.716  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag             10            N/A              New  thrpt   25      9043.742 ±     31.258  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag            100            N/A              Old  thrpt   25         0.383 ±      0.005  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag            100            N/A              New  thrpt   25       918.961 ±      3.304  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag            2            N/A              Old  thrpt   25    140646.234 ±   2049.645  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag            2            N/A              New  thrpt   25     66804.914 ±    554.233  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag           10            N/A              Old  thrpt   25     15182.335 ±    196.118  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag           10            N/A              New  thrpt   25     16246.895 ±    153.145  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag          100            N/A              Old  thrpt   25        43.990 ±      0.416  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag          100            N/A              New  thrpt   25      1615.993 ±      7.979  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag                2            N/A              Old  thrpt   25     35730.005 ±     95.861  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag                2            N/A              New  thrpt   25     38429.985 ±    339.620  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag               10            N/A              Old  thrpt   25      1089.862 ±      6.735  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag               10            N/A              New  thrpt   25      9829.611 ±    128.583  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag              100            N/A              Old  thrpt   25         0.325 ±      0.001  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag              100            N/A              New  thrpt   25       966.048 ±      4.876  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag              2            N/A              Old  thrpt   25     83371.705 ±    928.733  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag              2            N/A              New  thrpt   25     57001.385 ±    694.699  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag             10            N/A              Old  thrpt   25      3995.883 ±     42.895  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag             10            N/A              New  thrpt   25     15741.374 ±    182.879  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag            100            N/A              Old  thrpt   25         1.612 ±      0.001  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag            100            N/A              New  thrpt   25      1590.125 ±     11.210  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields             2            N/A              Old  thrpt   25    158147.729 ±   2360.497  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields             2            N/A              New  thrpt   25     74136.357 ±    435.569  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields            10            N/A              Old  thrpt   25     44564.947 ±    702.821  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields            10            N/A              New  thrpt   25     35614.600 ±    439.629  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields           100            N/A              Old  thrpt   25       343.071 ±     77.751  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields           100            N/A              New  thrpt   25      5227.893 ±     16.484  ops/s
StringUtilBenchmark.benchCurrent                                        N/A    empty input              N/A  thrpt   25   3174505.619 ± 165063.949  ops/s
StringUtilBenchmark.benchCurrent                                        N/A  empty options              N/A  thrpt   25  13509447.021 ±  28013.018  ops/s
StringUtilBenchmark.benchCurrent                                        N/A            abc              N/A  thrpt   25    391470.001 ±   5981.717  ops/s
StringUtilBenchmark.benchCurrent                                        N/A    description              N/A  thrpt   25     59222.403 ±    772.384  ops/s
StringUtilBenchmark.benchCurrent                                        N/A         length              N/A  thrpt   25      3992.155 ±    304.559  ops/s
```

### This Branch
```
Benchmark                                                            (size)     (testCase)  (validatorType)   Mode  Cnt         Score        Error  Units
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete       2            N/A              Old  thrpt   25     18104.287 ±     80.906  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete       2            N/A              New  thrpt   25     19093.390 ±    179.651  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete      10            N/A              Old  thrpt   25      5694.018 ±     76.354  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete      10            N/A              New  thrpt   25      6190.061 ±     39.913  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete     100            N/A              Old  thrpt   25       159.792 ±      1.249  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkDeepAbstractConcrete     100            N/A              New  thrpt   25       690.476 ±      3.925  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag              2            N/A              Old  thrpt   25     68645.227 ±    964.186  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag              2            N/A              New  thrpt   25     41582.757 ±    260.039  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag             10            N/A              Old  thrpt   25      2852.819 ±     22.037  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag             10            N/A              New  thrpt   25      9001.299 ±     63.872  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag            100            N/A              Old  thrpt   25         0.380 ±      0.005  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapFrag            100            N/A              New  thrpt   25       911.961 ±     10.497  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag            2            N/A              Old  thrpt   25    135292.766 ±   2410.502  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag            2            N/A              New  thrpt   25     65494.889 ±    271.182  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag           10            N/A              Old  thrpt   25     15324.461 ±    136.964  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag           10            N/A              New  thrpt   25     16158.985 ±    142.704  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag          100            N/A              Old  thrpt   25        45.034 ±      1.020  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkNoOverlapNoFrag          100            N/A              New  thrpt   25      1583.145 ±     11.358  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag                2            N/A              Old  thrpt   25     33781.361 ±    641.632  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag                2            N/A              New  thrpt   25     37518.901 ±    378.880  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag               10            N/A              Old  thrpt   25      1084.273 ±      6.260  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag               10            N/A              New  thrpt   25      9843.057 ±     72.879  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag              100            N/A              Old  thrpt   25         0.324 ±      0.001  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapFrag              100            N/A              New  thrpt   25       969.145 ±     11.045  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag              2            N/A              Old  thrpt   25     82607.314 ±   2301.080  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag              2            N/A              New  thrpt   25     57219.876 ±    307.253  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag             10            N/A              Old  thrpt   25      3963.680 ±     35.353  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag             10            N/A              New  thrpt   25     15759.511 ±     51.500  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag            100            N/A              Old  thrpt   25         1.608 ±      0.002  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkOverlapNoFrag            100            N/A              New  thrpt   25      1588.996 ±     13.400  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields             2            N/A              Old  thrpt   25    161431.527 ±   1168.393  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields             2            N/A              New  thrpt   25     74591.336 ±    615.911  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields            10            N/A              Old  thrpt   25     44976.498 ±    574.491  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields            10            N/A              New  thrpt   25     35580.583 ±    241.572  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields           100            N/A              Old  thrpt   25       458.454 ±      7.928  ops/s
OverlappingFieldsCanBeMergedBenchmark.benchmarkRepeatedFields           100            N/A              New  thrpt   25      5217.492 ±     22.090  ops/s
StringUtilBenchmark.benchCurrent                                        N/A    empty input              N/A  thrpt   25   2883757.741 ± 212449.441  ops/s
StringUtilBenchmark.benchCurrent                                        N/A  empty options              N/A  thrpt   25  12930546.372 ± 165473.653  ops/s
StringUtilBenchmark.benchCurrent                                        N/A            abc              N/A  thrpt   25    385983.074 ±   6068.929  ops/s
StringUtilBenchmark.benchCurrent                                        N/A    description              N/A  thrpt   25     56930.366 ±    512.665  ops/s
StringUtilBenchmark.benchCurrent                                        N/A         length              N/A  thrpt   25      4135.206 ±    315.175  ops/s

```